Rebase changes from Browser Provider PR
