+++ plugin/src/software/aws/toolkits/eclipse/amazonq/views/ToolkitLoginWebview.java
@@ -3,13 +3,6 @@
 
 package software.aws.toolkits.eclipse.amazonq.views;
 
-import java.io.IOException;
-import java.net.URL;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
 import org.eclipse.swt.browser.BrowserFunction;
 import org.eclipse.swt.browser.ProgressAdapter;
 import org.eclipse.swt.browser.ProgressEvent;
@@ -18,10 +11,9 @@ import org.eclipse.swt.widgets.Display;
 
 import software.aws.toolkits.eclipse.amazonq.lsp.auth.model.AuthState;
 import software.aws.toolkits.eclipse.amazonq.plugin.Activator;
+import software.aws.toolkits.eclipse.amazonq.providers.assets.ToolkitLoginWebViewAssetProvider;
+import software.aws.toolkits.eclipse.amazonq.providers.assets.WebViewAssetProvider;
 import software.aws.toolkits.eclipse.amazonq.telemetry.UiTelemetryProvider;
-import software.aws.toolkits.eclipse.amazonq.util.PluginUtils;
-import software.aws.toolkits.eclipse.amazonq.util.ThemeDetector;
-import software.aws.toolkits.eclipse.amazonq.util.WebviewAssetServer;
 import software.aws.toolkits.eclipse.amazonq.views.actions.AmazonQCommonActions;
 
 public final class ToolkitLoginWebview extends AmazonQView {
@@ -29,9 +21,8 @@ public final class ToolkitLoginWebview extends AmazonQView {
     public static final String ID = "software.aws.toolkits.eclipse.amazonq.views.ToolkitLoginWebview";
 
     private AmazonQCommonActions amazonQCommonActions;
-    private WebviewAssetServer webviewAssetServer;
-    private static final ThemeDetector THEME_DETECTOR = new ThemeDetector();
 
+    private final WebViewAssetProvider webViewAssetProvider;
     private final ViewCommandParser commandParser;
     private final ViewActionHandler actionHandler;
 
@@ -41,6 +32,7 @@ public final class ToolkitLoginWebview extends AmazonQView {
         super();
         this.commandParser = new LoginViewCommandParser();
         this.actionHandler = new LoginViewActionHandler();
+        this.webViewAssetProvider = new ToolkitLoginWebViewAssetProvider();
     }
 
     @Override
@@ -102,7 +94,7 @@ public final class ToolkitLoginWebview extends AmazonQView {
             amazonQCommonActions.updateActionVisibility(authState, getViewSite());
             if (!authState.isLoggedIn()) {
                 if (!browser.isDisposed()) {
-                    browser.setText(getContent());
+                    browser.setText(webViewAssetProvider.getContent().get());
                 }
             } else {
                 ViewVisibilityManager.showChatView("update");
@@ -110,82 +102,9 @@ public final class ToolkitLoginWebview extends AmazonQView {
         });
     }
 
-    private String getContent() {
-        try {
-            URL jsFile = PluginUtils.getResource("webview/build/assets/js/getStart.js");
-            String decodedPath = URLDecoder.decode(jsFile.getPath(), StandardCharsets.UTF_8);
-
-            // Remove leading slash for Windows paths
-            decodedPath = decodedPath.replaceFirst("^/([A-Za-z]:)", "$1");
-
-            Path jsParent = Paths.get(decodedPath).getParent();
-            String jsDirectoryPath = jsParent.normalize().toString();
-
-            webviewAssetServer = new WebviewAssetServer();
-            var result = webviewAssetServer.resolve(jsDirectoryPath);
-            if (!result) {
-                return "Failed to load JS";
-            }
-            var loginJsPath = webviewAssetServer.getUri() + "getStart.js";
-            boolean isDarkTheme = THEME_DETECTOR.isDarkTheme();
-            return String.format(
-                    """
-                            <!DOCTYPE html>
-                            <html>
-                                <head>
-                                    <meta
-                                        http-equiv="Content-Security-Policy"
-                                        content="default-src 'none'; script-src %s 'unsafe-inline'; style-src %s 'unsafe-inline';
-                                        img-src 'self' data:; object-src 'none'; base-uri 'none'; connect-src swt:;"
-                                    >
-                                    <title>AWS Q</title>
-                                </head>
-                                <body class="jb-light">
-                                    <div id="app"></div>
-                                    <script type="text/javascript" src="%s" defer></script>
-                                    <script type="text/javascript">
-                                        %s
-                                        const init = () => {
-                                            changeTheme(%b);
-                                            Promise.all([
-                                                waitForFunction('ideCommand'),
-                                                waitForFunction('telemetryEvent')
-                                            ])
-                                                .then(([ideCommand, telemetryEvent]) => {
-                                                    const ideApi = {
-                                                        postMessage(message) {
-                                                            ideCommand(JSON.stringify(message));
-                                                        }
-                                                    };
-                                                    window.ideApi = ideApi;
-
-                                                    const telemetryApi = {
-                                                        postClickEvent(event) {
-                                                            telemetryEvent(event);
-                                                        }
-                                                    };
-                                                    window.telemetryApi = telemetryApi;
-
-                                                    ideCommand(JSON.stringify({"command":"onLoad"}));
-                                                })
-                                                .catch(error => console.error('Error in initialization:', error));
-                                        };
-                                        window.addEventListener('load', init);
-                                    </script>
-                                </body>
-                            </html>
-                            """,
-                    loginJsPath, loginJsPath, loginJsPath, getWaitFunction(), isDarkTheme);
-        } catch (IOException e) {
-            return "Failed to load JS";
-        }
-    }
-
     @Override
     public void dispose() {
-        if (webviewAssetServer != null) {
-            webviewAssetServer.stop();
-        }
+        webViewAssetProvider.dispose();
         var browser = getBrowser();
         if (browser != null && !browser.isDisposed()) {
             browser.dispose();
