+++ plugin/src/software/aws/toolkits/eclipse/amazonq/views/router/PluginState.java
@@ -5,7 +5,9 @@ package software.aws.toolkits.eclipse.amazonq.views.router;
 
 import software.aws.toolkits.eclipse.amazonq.lsp.auth.model.AuthState;
 import software.aws.toolkits.eclipse.amazonq.lsp.manager.LspState;
+import software.aws.toolkits.eclipse.amazonq.providers.assets.WebViewAssetState;
 import software.aws.toolkits.eclipse.amazonq.providers.browser.BrowserCompatibilityState;
 
-public record PluginState(AuthState authState, LspState lspState, BrowserCompatibilityState browserCompatibilityState) {
+public record PluginState(AuthState authState, LspState lspState, BrowserCompatibilityState browserCompatibilityState,
+        WebViewAssetState webViewAssetState) {
 }
