+++ plugin/tst/software/aws/toolkits/eclipse/amazonq/views/router/ViewRouterTest.java
@@ -23,6 +23,7 @@ import software.aws.toolkits.eclipse.amazonq.extensions.implementation.Activator
 import software.aws.toolkits.eclipse.amazonq.lsp.auth.model.AuthState;
 import software.aws.toolkits.eclipse.amazonq.lsp.auth.model.AuthStateType;
 import software.aws.toolkits.eclipse.amazonq.lsp.manager.LspState;
+import software.aws.toolkits.eclipse.amazonq.providers.assets.WebViewAssetState;
 import software.aws.toolkits.eclipse.amazonq.providers.browser.BrowserCompatibilityState;
 
 public final class ViewRouterTest {
@@ -32,6 +33,7 @@ public final class ViewRouterTest {
     private Observable<AuthState> authStateObservable;
     private Observable<LspState> lspStateObservable;
     private Observable<BrowserCompatibilityState> browserCompatibilityStateObservable;
+    private Observable<WebViewAssetState> webViewAssetStateObservable;
 
     private ViewRouter viewRouter;
     private EventBroker eventBrokerMock;
@@ -49,12 +51,14 @@ public final class ViewRouterTest {
         authStateObservable = publishSubject.ofType(AuthState.class);
         lspStateObservable = publishSubject.ofType(LspState.class);
         browserCompatibilityStateObservable = publishSubject.ofType(BrowserCompatibilityState.class);
+        webViewAssetStateObservable = publishSubject.ofType(WebViewAssetState.class);
 
         eventBrokerMock = activatorStaticMockExtension.getMock(EventBroker.class);
 
         viewRouter = ViewRouter.builder().withAuthStateObservable(authStateObservable)
                 .withLspStateObservable(lspStateObservable)
-                .withBrowserCompatibilityStateObservable(browserCompatibilityStateObservable).build();
+                .withBrowserCompatibilityStateObservable(browserCompatibilityStateObservable)
+                .withWebViewAssetStateObservable(webViewAssetStateObservable).build();
     }
 
     @AfterEach
@@ -65,37 +69,47 @@ public final class ViewRouterTest {
     @ParameterizedTest
     @MethodSource("provideStateSource")
     void testActiveViewResolutionBasedOnPluginState(final LspState lspState, final AuthState authState,
-            final BrowserCompatibilityState browserCompatibilityState, final AmazonQViewType expectedActiveViewId) {
+            final BrowserCompatibilityState browserCompatibilityState, final WebViewAssetState webViewAssetState,
+            final AmazonQViewType expectedActiveViewId) {
         publishSubject.onNext(authState);
         publishSubject.onNext(lspState);
         publishSubject.onNext(browserCompatibilityState);
+        publishSubject.onNext(webViewAssetState);
 
         verify(eventBrokerMock).post(AmazonQViewType.class, expectedActiveViewId);
     }
 
     private static Stream<Arguments> provideStateSource() {
         return Stream.of(
+                Arguments.of(LspState.ACTIVE, getAuthStateObject(AuthStateType.LOGGED_IN),
+                        BrowserCompatibilityState.DEPENDENCY_MISSING, WebViewAssetState.RESOLVED,
+                        AmazonQViewType.DEPENDENCY_MISSING_VIEW),
+                Arguments.of(LspState.ACTIVE, getAuthStateObject(AuthStateType.LOGGED_IN),
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.DEPENDENCY_MISSING,
+                        AmazonQViewType.CHAT_ASSET_MISSING_VIEW),
                 Arguments.of(LspState.FAILED, getAuthStateObject(AuthStateType.LOGGED_IN),
-                        BrowserCompatibilityState.DEPENDENCY_MISSING, AmazonQViewType.DEPENDENCY_MISSING_VIEW),
-                Arguments.of(LspState.FAILED, getAuthStateObject(AuthStateType.LOGGED_IN),
-                        BrowserCompatibilityState.COMPATIBLE,
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
                         AmazonQViewType.LSP_STARTUP_FAILED_VIEW),
                 Arguments.of(LspState.FAILED, getAuthStateObject(AuthStateType.LOGGED_OUT),
-                        BrowserCompatibilityState.COMPATIBLE,
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
                         AmazonQViewType.LSP_STARTUP_FAILED_VIEW),
                 Arguments.of(LspState.FAILED, getAuthStateObject(AuthStateType.EXPIRED),
-                        BrowserCompatibilityState.COMPATIBLE,
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
                         AmazonQViewType.LSP_STARTUP_FAILED_VIEW),
                 Arguments.of(LspState.PENDING, getAuthStateObject(AuthStateType.LOGGED_OUT),
-                        BrowserCompatibilityState.COMPATIBLE, AmazonQViewType.TOOLKIT_LOGIN_VIEW),
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
+                        AmazonQViewType.TOOLKIT_LOGIN_VIEW),
                 Arguments.of(LspState.ACTIVE, getAuthStateObject(AuthStateType.LOGGED_OUT),
-                        BrowserCompatibilityState.COMPATIBLE, AmazonQViewType.TOOLKIT_LOGIN_VIEW),
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
+                        AmazonQViewType.TOOLKIT_LOGIN_VIEW),
                 Arguments.of(LspState.PENDING, getAuthStateObject(AuthStateType.EXPIRED),
-                        BrowserCompatibilityState.COMPATIBLE, AmazonQViewType.RE_AUTHENTICATE_VIEW),
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
+                        AmazonQViewType.RE_AUTHENTICATE_VIEW),
                 Arguments.of(LspState.ACTIVE, getAuthStateObject(AuthStateType.EXPIRED),
-                        BrowserCompatibilityState.COMPATIBLE, AmazonQViewType.RE_AUTHENTICATE_VIEW),
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED,
+                        AmazonQViewType.RE_AUTHENTICATE_VIEW),
                 Arguments.of(LspState.ACTIVE, getAuthStateObject(AuthStateType.LOGGED_IN),
-                        BrowserCompatibilityState.COMPATIBLE, AmazonQViewType.CHAT_VIEW));
+                        BrowserCompatibilityState.COMPATIBLE, WebViewAssetState.RESOLVED, AmazonQViewType.CHAT_VIEW));
     }
 
     private static AuthState getAuthStateObject(final AuthStateType authStateType) {
